+++ program/src/processor.rs
@@ -5,7 +5,6 @@ use std::mem::size_of;
 use std::vec;
 
 use arrayref::{array_ref, array_refs};
-use borsh::BorshSerialize;
 use bytemuck::{cast, cast_mut, cast_ref};
 use fixed::types::I80F48;
 use serum_dex::instruction::NewOrderInstructionV3;
@@ -14,7 +13,6 @@ use solana_program::account_info::AccountInfo;
 use solana_program::clock::Clock;
 use solana_program::entrypoint::ProgramResult;
 use solana_program::instruction::{AccountMeta, Instruction};
-use solana_program::log::sol_log_compute_units;
 use solana_program::msg;
 use solana_program::program_error::ProgramError;
 use solana_program::program_pack::{IsInitialized, Pack};
@@ -32,7 +30,7 @@ use crate::ids::srm_token;
 use crate::instruction::MangoInstruction;
 use crate::matching::{Book, BookSide, OrderType, Side};
 use crate::oracle::{determine_oracle_type, OracleType, Price, StubOracle};
-use crate::queue::{EventQueue, EventType, FillEvent, LiquidateEvent, LoggableFillEvent, OutEvent};
+use crate::queue::{EventQueue, EventType, FillEvent, LiquidateEvent, OutEvent};
 use crate::state::{
     load_asks_mut, load_bids_mut, load_market_state, load_open_orders, AdvancedOrderType,
     AdvancedOrders, AssetType, DataType, HealthCache, HealthType, MangoAccount, MangoCache,
@@ -42,6 +40,13 @@ use crate::state::{
     MAX_NODE_BANKS, MAX_PAIRS, MAX_PERP_OPEN_ORDERS, ONE_I80F48, QUOTE_INDEX, ZERO_I80F48,
 };
 use crate::utils::{gen_signer_key, gen_signer_seeds};
+use anchor_lang::prelude::emit;
+use mango_logs::{
+    mango_emit, CachePerpMarketsLog, CachePricesLog, CacheRootBanksLog, LiquidatePerpMarketLog,
+    LiquidateTokenAndPerpLog, LiquidateTokenAndTokenLog, MngoAccrualLog, OpenOrdersBalanceLog,
+    PerpBankruptcyLog, SettleFeesLog, SettlePnlLog, TokenBalanceLog, TokenBankruptcyLog,
+    UpdateRootBankLog,
+};
 
 declare_check_assert_macros!(SourceFileId::Processor);
 
@@ -270,12 +275,15 @@ impl Processor {
             padding: [0u8; 7],
         };
 
+        let (maint_asset_weight, maint_liab_weight) = get_leverage_weights(maint_leverage);
+        let (init_asset_weight, init_liab_weight) = get_leverage_weights(init_leverage);
+
         mango_group.spot_markets[market_index] = SpotMarketInfo {
             spot_market: *spot_market_ai.key,
-            maint_asset_weight: (maint_leverage - ONE_I80F48).checked_div(maint_leverage).unwrap(),
-            init_asset_weight: (init_leverage - ONE_I80F48).checked_div(init_leverage).unwrap(),
-            maint_liab_weight: (maint_leverage + ONE_I80F48).checked_div(maint_leverage).unwrap(),
-            init_liab_weight: (init_leverage + ONE_I80F48).checked_div(init_leverage).unwrap(),
+            maint_asset_weight,
+            init_asset_weight,
+            maint_liab_weight,
+            init_liab_weight,
             liquidation_fee,
         };
 
@@ -316,8 +324,8 @@ impl Processor {
         ] = accounts;
 
         let mut mango_group = MangoGroup::load_mut_checked(mango_group_ai, program_id)?;
-        check!(admin_ai.is_signer, MangoErrorCode::InvalidSignerKey)?;
-        check_eq!(admin_ai.key, &mango_group.admin, MangoErrorCode::InvalidSignerKey)?;
+        check!(admin_ai.is_signer, MangoErrorCode::SignerNecessary)?;
+        check_eq!(admin_ai.key, &mango_group.admin, MangoErrorCode::InvalidAdminKey)?;
 
         let oracle_type = determine_oracle_type(oracle_ai);
         match oracle_type {
@@ -353,8 +361,8 @@ impl Processor {
         ] = accounts;
 
         let mango_group = MangoGroup::load_checked(mango_group_ai, program_id)?;
-        check!(admin_ai.is_signer, MangoErrorCode::Default)?;
-        check_eq!(admin_ai.key, &mango_group.admin, MangoErrorCode::Default)?;
+        check!(admin_ai.is_signer, MangoErrorCode::SignerNecessary)?;
+        check_eq!(admin_ai.key, &mango_group.admin, MangoErrorCode::InvalidAdminKey)?;
         check!(mango_group.find_oracle_index(oracle_ai.key).is_some(), MangoErrorCode::Default)?;
 
         let oracle_type = determine_oracle_type(oracle_ai);
@@ -425,12 +433,15 @@ impl Processor {
         // Make sure perp market at this index not already initialized
         check!(mango_group.perp_markets[market_index].is_empty(), MangoErrorCode::InvalidParam)?;
 
+        let (maint_asset_weight, maint_liab_weight) = get_leverage_weights(maint_leverage);
+        let (init_asset_weight, init_liab_weight) = get_leverage_weights(init_leverage);
+
         mango_group.perp_markets[market_index] = PerpMarketInfo {
             perp_market: *perp_market_ai.key,
-            maint_asset_weight: (maint_leverage - ONE_I80F48).checked_div(maint_leverage).unwrap(),
-            init_asset_weight: (init_leverage - ONE_I80F48).checked_div(init_leverage).unwrap(),
-            maint_liab_weight: (maint_leverage + ONE_I80F48).checked_div(maint_leverage).unwrap(),
-            init_liab_weight: (init_leverage + ONE_I80F48).checked_div(init_leverage).unwrap(),
+            maint_asset_weight,
+            init_asset_weight,
+            maint_liab_weight,
+            init_liab_weight,
             liquidation_fee,
             maker_fee,
             taker_fee,
@@ -558,7 +569,6 @@ impl Processor {
     }
     #[inline(never)]
     /// Change leverage, fees and liquidity mining params
-    /// Note: only mngo_per_period  and rate works right now; others not yet implemented; this is for safety
     fn change_perp_market_params(
         program_id: &Pubkey,
         accounts: &[AccountInfo],
@@ -596,12 +606,12 @@ impl Processor {
 
         // Unwrap params. Default to current state if Option is None
         let (maint_asset_weight, maint_liab_weight) = if let Some(x) = maint_leverage {
-            ((x - ONE_I80F48).checked_div(x).unwrap(), (x + ONE_I80F48).checked_div(x).unwrap())
+            get_leverage_weights(x)
         } else {
             (info.maint_asset_weight, info.maint_liab_weight)
         };
         let (init_asset_weight, init_liab_weight) = if let Some(x) = init_leverage {
-            ((x - ONE_I80F48).checked_div(x).unwrap(), (x + ONE_I80F48).checked_div(x).unwrap())
+            get_leverage_weights(x)
         } else {
             (info.init_asset_weight, info.init_liab_weight)
         };
@@ -680,7 +690,6 @@ impl Processor {
 
         let mut oracle_indexes = Vec::new();
         let mut oracle_prices = Vec::new();
-
         for oracle_ai in oracle_ais.iter() {
             let oracle_index = mango_group.find_oracle_index(oracle_ai.key).ok_or(throw!())?;
             let oracle_price = read_oracle(&mango_group, oracle_index, oracle_ai)?;
@@ -688,20 +697,15 @@ impl Processor {
             mango_cache.price_cache[oracle_index] =
                 PriceCache { price: oracle_price, last_update: now_ts };
 
-            oracle_indexes.push(oracle_index);
-            oracle_prices.push(oracle_price.to_num::<f64>());
+            oracle_indexes.push(oracle_index as u64);
+            oracle_prices.push(oracle_price.to_bits());
         }
 
-        msg!(
-            "cache_prices details: {{ \
-            \"mango_group_pk\": \"{}\", \
-            \"oracle_indexes\": {:?}, \
-            \"oracle_prices\": {:?} \
-        }}",
-            mango_group_ai.key,
+        mango_emit!(CachePricesLog {
+            mango_group: *mango_group_ai.key,
             oracle_indexes,
             oracle_prices
-        );
+        });
 
         Ok(())
     }
@@ -734,25 +738,16 @@ impl Processor {
                 last_update: now_ts,
             };
 
-            token_indexes.push(index);
-            deposit_indexes.push(root_bank.deposit_index.to_num::<f64>());
-            borrow_indexes.push(root_bank.borrow_index.to_num::<f64>())
+            token_indexes.push(index as u64);
+            deposit_indexes.push(root_bank.deposit_index.to_bits());
+            borrow_indexes.push(root_bank.borrow_index.to_bits())
         }
-
-        msg!(
-            "cache_root_banks details: {{ \
-            \"mango_group_pk\": \"{}\", \
-            \"token_indexes\": {:?}, \
-            \"deposit_indexes\": {:?}, \
-            \"borrow_indexes\": {:?}, \
-            \"redundant_field\": {}
-        }}",
-            mango_group_ai.key,
+        mango_emit!(CacheRootBanksLog {
+            mango_group: *mango_group_ai.key,
             token_indexes,
             deposit_indexes,
-            borrow_indexes,
-            0
-        );
+            borrow_indexes
+        });
 
         Ok(())
     }
@@ -786,25 +781,17 @@ impl Processor {
                 last_update: now_ts,
             };
 
-            market_indexes.push(index);
-            long_fundings.push(perp_market.long_funding.to_num::<f64>());
-            short_fundings.push(perp_market.short_funding.to_num::<f64>());
+            market_indexes.push(index as u64);
+            long_fundings.push(perp_market.long_funding.to_bits());
+            short_fundings.push(perp_market.short_funding.to_bits());
         }
 
-        msg!(
-            "cache_perp_markets details: {{ \
-                \"mango_group_pk\": \"{}\", \
-                \"market_indexes\": {:?}, \
-                \"long_funding\": {:?}, \
-                \"short_funding\": {:?}, \
-                \"redundant_field\": {} \
-                }}",
-            mango_group_ai.key,
+        mango_emit!(CachePerpMarketsLog {
+            mango_group: *mango_group_ai.key,
             market_indexes,
             long_fundings,
-            short_fundings,
-            0
-        );
+            short_fundings
+        });
 
         Ok(())
     }
@@ -1166,7 +1153,20 @@ impl Processor {
         check!(
             post_health >= ZERO_I80F48 || (reduce_only && post_health >= pre_health),
             MangoErrorCode::InsufficientFunds
-        )
+        )?;
+
+        mango_emit!(OpenOrdersBalanceLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            base_total: open_orders.native_coin_total,
+            base_free: open_orders.native_coin_free,
+            quote_total: open_orders.native_pc_total,
+            quote_free: open_orders.native_pc_free,
+            referrer_rebates_accrued: open_orders.referrer_rebates_accrued
+        });
+
+        Ok(())
     }
 
     #[inline(never)]
@@ -1338,9 +1338,6 @@ impl Processor {
 
         // See if we can remove this market from margin
         let open_orders = load_open_orders(open_orders_ais[market_index].unwrap())?;
-
-        // TODO - log what's in open orders
-
         mango_account.update_basket(market_index, &open_orders)?;
 
         let (post_base, post_quote) = {
@@ -1386,7 +1383,20 @@ impl Processor {
         check!(
             post_health >= ZERO_I80F48 || (reduce_only && post_health >= pre_health),
             MangoErrorCode::InsufficientFunds
-        )
+        )?;
+
+        mango_emit!(OpenOrdersBalanceLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            base_total: open_orders.native_coin_total,
+            base_free: open_orders.native_coin_free,
+            quote_total: open_orders.native_pc_total,
+            quote_free: open_orders.native_pc_free,
+            referrer_rebates_accrued: open_orders.referrer_rebates_accrued
+        });
+
+        Ok(())
     }
 
     #[inline(never)]
@@ -1414,19 +1424,19 @@ impl Processor {
         ] = accounts;
 
         let mango_group = MangoGroup::load_checked(mango_group_ai, program_id)?;
-        check_eq!(dex_prog_ai.key, &mango_group.dex_program_id, MangoErrorCode::Default)?;
+        check_eq!(dex_prog_ai.key, &mango_group.dex_program_id, MangoErrorCode::InvalidProgramId)?;
 
         let mango_account =
             MangoAccount::load_checked(mango_account_ai, program_id, mango_group_ai.key)?;
+        check!(&mango_account.owner == owner_ai.key, MangoErrorCode::InvalidOwner)?;
+        check!(owner_ai.is_signer, MangoErrorCode::InvalidSignerKey)?;
         check!(!mango_account.is_bankrupt, MangoErrorCode::Bankrupt)?;
-        check!(owner_ai.is_signer, MangoErrorCode::Default)?;
-        check_eq!(&mango_account.owner, owner_ai.key, MangoErrorCode::Default)?;
 
-        let market_i = mango_group.find_spot_market_index(spot_market_ai.key).unwrap();
+        let market_index = mango_group.find_spot_market_index(spot_market_ai.key).unwrap();
         check_eq!(
-            &mango_account.spot_open_orders[market_i],
+            &mango_account.spot_open_orders[market_index],
             open_orders_ai.key,
-            MangoErrorCode::Default
+            MangoErrorCode::InvalidOpenOrdersAccount
         )?;
 
         let signer_seeds = gen_signer_seeds(&mango_group.signer_nonce, mango_group_ai.key);
@@ -1441,6 +1451,19 @@ impl Processor {
             data,
             &[&signer_seeds],
         )?;
+
+        let open_orders = load_open_orders(open_orders_ai)?;
+        mango_emit!(OpenOrdersBalanceLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            base_total: open_orders.native_coin_total,
+            base_free: open_orders.native_coin_free,
+            quote_total: open_orders.native_pc_total,
+            quote_free: open_orders.native_pc_free,
+            referrer_rebates_accrued: open_orders.referrer_rebates_accrued
+        });
+
         Ok(())
     }
 
@@ -1470,8 +1493,8 @@ impl Processor {
         ] = accounts;
 
         let mango_group = MangoGroup::load_checked(mango_group_ai, program_id)?;
-        check_eq!(token_prog_ai.key, &spl_token::id(), MangoErrorCode::Default)?;
-        check_eq!(dex_prog_ai.key, &mango_group.dex_program_id, MangoErrorCode::Default)?;
+        check_eq!(token_prog_ai.key, &spl_token::id(), MangoErrorCode::InvalidProgramId)?;
+        check_eq!(dex_prog_ai.key, &mango_group.dex_program_id, MangoErrorCode::InvalidProgramId)?;
 
         let mut mango_account =
             MangoAccount::load_mut_checked(mango_account_ai, program_id, mango_group_ai.key)?;
@@ -1480,13 +1503,13 @@ impl Processor {
         check!(!mango_account.is_bankrupt, MangoErrorCode::Bankrupt)?;
 
         // Make sure the spot market is valid
-        let spot_market_index = mango_group
+        let market_index = mango_group
             .find_spot_market_index(spot_market_ai.key)
             .ok_or(throw_err!(MangoErrorCode::InvalidMarket))?;
 
         let base_root_bank = RootBank::load_checked(base_root_bank_ai, program_id)?;
         check!(
-            base_root_bank_ai.key == &mango_group.tokens[spot_market_index].root_bank,
+            base_root_bank_ai.key == &mango_group.tokens[market_index].root_bank,
             MangoErrorCode::InvalidRootBank
         )?;
 
@@ -1510,7 +1533,7 @@ impl Processor {
         check_eq!(&quote_node_bank.vault, quote_vault_ai.key, MangoErrorCode::InvalidVault)?;
 
         check_eq!(
-            &mango_account.spot_open_orders[spot_market_index],
+            &mango_account.spot_open_orders[market_index],
             open_orders_ai.key,
             MangoErrorCode::Default
         )?;
@@ -1545,7 +1568,17 @@ impl Processor {
         let (post_base, post_quote) = {
             let open_orders = load_open_orders(open_orders_ai)?;
             // remove from margin basket if it's empty
-            mango_account.update_basket(spot_market_index, &open_orders)?;
+            mango_account.update_basket(market_index, &open_orders)?;
+            mango_emit!(OpenOrdersBalanceLog {
+                mango_group: *mango_group_ai.key,
+                mango_account: *mango_account_ai.key,
+                market_index: market_index as u64,
+                base_total: open_orders.native_coin_total,
+                base_free: open_orders.native_coin_free,
+                quote_total: open_orders.native_pc_total,
+                quote_free: open_orders.native_pc_free,
+                referrer_rebates_accrued: open_orders.referrer_rebates_accrued
+            });
 
             (
                 open_orders.native_coin_free,
@@ -1554,13 +1587,13 @@ impl Processor {
         };
 
         // TODO OPT - remove sanity check if confident
-        check!(post_base <= pre_base, MangoErrorCode::Default)?;
-        check!(post_quote <= pre_quote, MangoErrorCode::Default)?;
+        check!(post_base <= pre_base, MangoErrorCode::MathError)?;
+        check!(post_quote <= pre_quote, MangoErrorCode::MathError)?;
 
         let mango_cache = MangoCache::load_checked(mango_cache_ai, program_id, &mango_group)?;
 
         let now_ts = Clock::get()?.unix_timestamp as u64;
-        let base_root_bank_cache = &mango_cache.root_bank_cache[spot_market_index];
+        let base_root_bank_cache = &mango_cache.root_bank_cache[market_index];
         let quote_root_bank_cache = &mango_cache.root_bank_cache[QUOTE_INDEX];
 
         base_root_bank_cache.check_valid(&mango_group, now_ts)?;
@@ -1571,7 +1604,7 @@ impl Processor {
             &mut base_node_bank,
             &mut mango_account,
             mango_account_ai.key,
-            spot_market_index,
+            market_index,
             I80F48::from_num(pre_base - post_base),
         )?;
         checked_change_net(
@@ -1721,7 +1754,7 @@ impl Processor {
         let mut mango_account =
             MangoAccount::load_mut_checked(mango_account_ai, program_id, mango_group_ai.key)?;
         check!(!mango_account.is_bankrupt, MangoErrorCode::Bankrupt)?;
-        check!(owner_ai.is_signer, MangoErrorCode::Default)?;
+        check!(owner_ai.is_signer, MangoErrorCode::SignerNecessary)?;
         check_eq!(&mango_account.owner, owner_ai.key, MangoErrorCode::InvalidOwner)?;
 
         let mut perp_market =
@@ -1745,6 +1778,7 @@ impl Processor {
         mango_account.remove_order(order.owner_slot as usize, order.quantity)?;
 
         let perp_account = &mut mango_account.perp_accounts[market_index];
+        let mngo_start = perp_account.mngo_accrued;
         perp_account.apply_incentives(
             &mut perp_market,
             side,
@@ -1755,6 +1789,12 @@ impl Processor {
             Clock::get()?.unix_timestamp as u64,
             order.quantity,
         )?;
+        mango_emit!(MngoAccrualLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            mngo_accrual: perp_account.mngo_accrued - mngo_start
+        });
 
         Ok(())
     }
@@ -1784,7 +1824,7 @@ impl Processor {
         let mut mango_account =
             MangoAccount::load_mut_checked(mango_account_ai, program_id, mango_group_ai.key)?;
         check!(!mango_account.is_bankrupt, MangoErrorCode::Bankrupt)?;
-        check!(owner_ai.is_signer, MangoErrorCode::Default)?;
+        check!(owner_ai.is_signer, MangoErrorCode::SignerNecessary)?;
         check_eq!(&mango_account.owner, owner_ai.key, MangoErrorCode::InvalidOwner)?;
 
         let mut perp_market =
@@ -1804,6 +1844,7 @@ impl Processor {
         let order = book.cancel_order(order_id, side)?;
         check_eq!(&order.owner, mango_account_ai.key, MangoErrorCode::InvalidOrderId)?;
         mango_account.remove_order(order.owner_slot as usize, order.quantity)?;
+        let mngo_start = mango_account.perp_accounts[market_index].mngo_accrued;
         mango_account.perp_accounts[market_index].apply_incentives(
             &mut perp_market,
             side,
@@ -1814,6 +1855,12 @@ impl Processor {
             Clock::get()?.unix_timestamp as u64,
             order.quantity,
         )?;
+        mango_emit!(MngoAccrualLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            mngo_accrual: mango_account.perp_accounts[market_index].mngo_accrued - mngo_start
+        });
 
         Ok(())
     }
@@ -1849,7 +1896,15 @@ impl Processor {
         let market_index = mango_group.find_perp_market_index(perp_market_ai.key).unwrap();
 
         let mut book = Book::load_checked(program_id, bids_ai, asks_ai, &perp_market)?;
-        book.cancel_all_with_incentives(&mut mango_account, &mut perp_market, market_index, limit)
+        let mngo_start = mango_account.perp_accounts[market_index].mngo_accrued;
+        book.cancel_all_with_incentives(&mut mango_account, &mut perp_market, market_index, limit)?;
+        mango_emit!(MngoAccrualLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            mngo_accrual: mango_account.perp_accounts[market_index].mngo_accrued - mngo_start
+        });
+        Ok(())
     }
 
     #[inline(never)]
@@ -1939,20 +1994,13 @@ impl Processor {
             a_settle,
         )?;
 
-        msg!(
-            "settle_pnl details: {{ \
-                \"mango_group_pk\": \"{}\", \
-                \"mango_account_a_pk\": \"{}\", \
-                \"mango_account_b_pk\": \"{}\", \
-                \"market_index\": {}, \
-                \"settlement\": {}
-                }}",
-            mango_group_ai.key,
-            mango_account_a_ai.key,
-            mango_account_b_ai.key,
-            market_index,
-            a_settle.to_num::<f64>() // Will be positive if a has positive pnl and settling with b
-        );
+        mango_emit!(SettlePnlLog {
+            mango_group: *mango_group_ai.key,
+            mango_account_a: *mango_account_a_ai.key,
+            mango_account_b: *mango_account_b_ai.key,
+            market_index: market_index as u64,
+            settlement: a_settle.to_bits(), // Will be positive if a has positive pnl and settling with b
+        });
 
         Ok(())
     }
@@ -2043,18 +2091,12 @@ impl Processor {
             -settlement,
         )?;
 
-        msg!(
-            "settle_fees details: {{ \
-                \"mango_group_pk\": \"{}\", \
-                \"mango_account_pk\": \"{}\", \
-                \"market_index\": {}, \
-                \"settlement\": {} \
-                }}",
-            mango_group_ai.key,
-            mango_account_ai.key,
-            market_index,
-            settlement.to_num::<f64>(), // Will be positive if a has positive pnl and settling with b
-        );
+        mango_emit!(SettleFeesLog {
+            mango_group: *mango_group_ai.key,
+            mango_account: *mango_account_ai.key,
+            market_index: market_index as u64,
+            settlement: settlement.to_bits()
+        });
 
         Ok(())
     }
@@ -2205,6 +2247,17 @@ impl Processor {
         let (post_base, post_quote) = {
             let open_orders = load_open_orders(open_orders_ai)?;
             liqee_ma.update_basket(market_index, &open_orders)?;
+            mango_emit!(OpenOrdersBalanceLog {
+                mango_group: *mango_group_ai.key,
+                mango_account: *liqee_mango_account_ai.key,
+                market_index: market_index as u64,
+                base_total: open_orders.native_coin_total,
+                base_free: open_orders.native_coin_free,
+                quote_total: open_orders.native_pc_total,
+                quote_free: open_orders.native_pc_free,
+                referrer_rebates_accrued: open_orders.referrer_rebates_accrued
+            });
+
             (
                 open_orders.native_coin_free,
                 open_orders.native_pc_free + open_orders.referrer_rebates_accrued,
@@ -2509,30 +2562,18 @@ impl Processor {
             }
         }
 
-        msg!(
-            "liquidate_token_and_token details: {{ \
-            \"mango_group_pk\": \"{}\", \
-            \"liqee_pk\": \"{}\", \
-            \"liqor_pk\": \"{}\", \
-            \"asset_index\": {}, \
-            \"liab_index\": {}, \
-            \"asset_transfer\": {}, \
-            \"liab_transfer\": {}, \
-            \"asset_price\": {}, \
-            \"liab_price\": {}, \
-            \"bankruptcy\": {} \
-        }}",
-            mango_group_ai.key,
-            liqee_mango_account_ai.key,
-            liqor_mango_account_ai.key,
-            asset_index,
-            liab_index,
-            asset_transfer.to_num::<f64>(),
-            actual_liab_transfer.to_num::<f64>(),
-            asset_price.to_num::<f64>(),
-            liab_price.to_num::<f64>(),
-            liqee_ma.is_bankrupt
-        );
+        mango_emit!(LiquidateTokenAndTokenLog {
+            mango_group: *mango_group_ai.key,
+            liqee: *liqee_mango_account_ai.key,
+            liqor: *liqor_mango_account_ai.key,
+            asset_index: asset_index as u64,
+            liab_index: liab_index as u64,
+            asset_transfer: asset_transfer.to_bits(),
+            liab_transfer: actual_liab_transfer.to_bits(),
+            asset_price: asset_price.to_bits(),
+            liab_price: liab_price.to_bits(),
+            bankruptcy: liqee_ma.is_bankrupt
+        });
 
         Ok(())
     }
@@ -2805,32 +2846,20 @@ impl Processor {
             liqee_ma.being_liquidated = liqee_init_health < ZERO_I80F48;
         }
 
-        msg!(
-            "liquidate_token_and_perp details: {{ \
-            \"mango_group_pk\": \"{}\", \
-            \"liqee_pk\": \"{}\", \
-            \"liqor_pk\": \"{}\", \
-            \"asset_index\": {}, \
-            \"liab_index\": {}, \
-            \"asset_type\": \"{:?}\", \
-            \"liab_type\": \"{:?}\", \
-            \"asset_price\": {}, \
-            \"liab_price\": {}, \
-            \"asset_transfer\": {}, \
-            \"actual_liab_transfer\": {} \
-        }}",
-            mango_group_ai.key,
-            liqee_mango_account_ai.key,
-            liqor_mango_account_ai.key,
-            asset_index,
-            liab_index,
-            asset_type,
-            liab_type,
-            asset_price.to_num::<f64>(),
-            liab_price.to_num::<f64>(),
-            asset_transfer.to_num::<f64>(),
-            actual_liab_transfer.to_num::<f64>()
-        );
+        mango_emit!(LiquidateTokenAndPerpLog {
+            mango_group: *mango_group_ai.key,
+            liqee: *liqee_mango_account_ai.key,
+            liqor: *liqor_mango_account_ai.key,
+            asset_index: asset_index as u64,
+            liab_index: liab_index as u64,
+            asset_type: asset_type as u8,
+            liab_type: liab_type as u8,
+            asset_transfer: asset_transfer.to_bits(),
+            liab_transfer: actual_liab_transfer.to_bits(),
+            asset_price: asset_price.to_bits(),
+            liab_price: liab_price.to_bits(),
+            bankruptcy: liqee_ma.is_bankrupt,
+        });
 
         Ok(())
     }
@@ -3022,27 +3051,17 @@ impl Processor {
             liqee_ma.being_liquidated = liqee_init_health < ZERO_I80F48;
         }
 
-        // TODO OPT make this more efficient
-        msg!(
-            "liquidate_perp_market details: {{ \
-                \"mango_group_pk\": \"{}\", \
-                \"liqee_pk\": \"{}\", \
-                \"liqor_pk\": \"{}\", \
-                \"market_index\": {}, \
-                \"price\": {}, \
-                \"base_transfer\": {}, \
-                \"quote_transfer\": {}, \
-                \"bankruptcy\": {}
-            }}",
-            mango_group_ai.key,
-            liqee_mango_account_ai.key,
-            liqor_mango_account_ai.key,
-            market_index,
-            price,
+        mango_emit!(LiquidatePerpMarketLog {
+            mango_group: *mango_group_ai.key,
+            liqee: *liqee_mango_account_ai.key,
+            liqor: *liqor_mango_account_ai.key,
+            market_index: market_index as u64,
+            price: price.to_bits(),
             base_transfer,
-            quote_transfer.to_num::<f64>(),
-            liqee_ma.is_bankrupt,
-        );
+            quote_transfer: quote_transfer.to_bits(),
+            bankruptcy: liqee_ma.is_bankrupt
+        });
+
         Ok(())
     }
 
@@ -3167,58 +3186,41 @@ impl Processor {
 
             let liqor_health = liqor_health_cache.get_health(&mango_group, HealthType::Init);
             check!(liqor_health >= ZERO_I80F48, MangoErrorCode::InsufficientFunds)?;
-            msg!(
-                "perp_bankruptcy details: {{ \
-                    \"mango_group_pk\": \"{}\", \
-                    \"liqee_pk\": \"{}\", \
-                    \"liqor_pk\": \"{}\", \
-                    \"liab_index\": {}, \
-                    \"insurance_transfer\": {} \
-                }}",
-                mango_group_ai.key,
-                liqee_mango_account_ai.key,
-                liqor_mango_account_ai.key,
-                liab_index,
-                liab_transfer_u64
-            );
         }
 
         let quote_position = liqee_ma.perp_accounts[liab_index].quote_position;
         // If we transferred everything out of insurance_vault, insurance vault is empty
         // and if quote position is still negative
-        if liab_transfer_u64 == insurance_vault.amount && quote_position.is_negative() {
-            // insurance fund empty so socialize loss
-            check!(
-                &mango_group.perp_markets[liab_index].perp_market == perp_market_ai.key,
-                MangoErrorCode::InvalidMarket
-            )?;
-            let mut perp_market =
-                PerpMarket::load_mut_checked(perp_market_ai, program_id, mango_group_ai.key)?;
+        let socialized_loss =
+            if liab_transfer_u64 == insurance_vault.amount && quote_position.is_negative() {
+                // insurance fund empty so socialize loss
+                check!(
+                    &mango_group.perp_markets[liab_index].perp_market == perp_market_ai.key,
+                    MangoErrorCode::InvalidMarket
+                )?;
+                let mut perp_market =
+                    PerpMarket::load_mut_checked(perp_market_ai, program_id, mango_group_ai.key)?;
 
-            perp_market.socialize_loss(
-                &mut liqee_ma.perp_accounts[liab_index],
-                &mut mango_cache.perp_market_cache[liab_index],
-            )?;
-            // TODO - log the cached long funding and short funding
-            // TODO test compute limits on devnet
-
-            msg!(
-                "perp_socialized_loss details: {{ \
-                    \"mango_group_pk\": \"{}\", \
-                    \"liqee_pk\": \"{}\", \
-                    \"liqor_pk\": \"{}\", \
-                    \"liab_index\": {}, \
-                    \"socialized_loss\":{} \
-                }}",
-                mango_group_ai.key,
-                liqee_mango_account_ai.key,
-                liqor_mango_account_ai.key,
-                liab_index,
-                (quote_position / (I80F48::from_num(perp_market.open_interest))).to_num::<f64>()
-            );
-        }
+                perp_market.socialize_loss(
+                    &mut liqee_ma.perp_accounts[liab_index],
+                    &mut mango_cache.perp_market_cache[liab_index],
+                )?
+            } else {
+                ZERO_I80F48
+            };
 
         liqee_ma.is_bankrupt = !liqee_ma.check_exit_bankruptcy(&mango_group);
+
+        mango_emit!(PerpBankruptcyLog {
+            mango_group: *mango_group_ai.key,
+            liqee: *liqee_mango_account_ai.key,
+            liqor: *liqor_mango_account_ai.key,
+            liab_index: liab_index as u64,
+            insurance_transfer: liab_transfer_u64,
+            socialized_loss: socialized_loss.to_bits(),
+            cache_long_funding: mango_cache.perp_market_cache[liab_index].long_funding.to_bits(),
+            cache_short_funding: mango_cache.perp_market_cache[liab_index].short_funding.to_bits()
+        });
 
         Ok(())
     }
@@ -3393,23 +3395,9 @@ impl Processor {
             )?;
             let liqor_health = liqor_health_cache.get_health(&mango_group, HealthType::Init);
             check!(liqor_health >= ZERO_I80F48, MangoErrorCode::InsufficientFunds)?;
-            msg!(
-                "token_bankruptcy details: {{ \
-                    \"mango_group_pk\": \"{}\", \
-                    \"liqee_pk\": \"{}\", \
-                    \"liqor_pk\": \"{}\", \
-                    \"liab_index\": {}, \
-                    \"insurance_transfer\": {} \
-                }}",
-                mango_group_ai.key,
-                liqee_mango_account_ai.key,
-                liqor_mango_account_ai.key,
-                liab_index,
-                insurance_transfer
-            );
         }
 
-        if insurance_transfer == insurance_vault.amount
+        let (socialized_loss, percentage_loss) = if insurance_transfer == insurance_vault.amount
             && liqee_ma.borrows[liab_index].is_positive()
         {
             // insurance fund empty so socialize loss
@@ -3419,10 +3407,23 @@ impl Processor {
                 &mut mango_cache,
                 &mut liqee_ma,
                 liab_node_bank_ais,
-            )?;
-        }
+            )?
+        } else {
+            (ZERO_I80F48, ZERO_I80F48)
+        };
 
         liqee_ma.is_bankrupt = !liqee_ma.check_exit_bankruptcy(&mango_group);
+
+        mango_emit!(TokenBankruptcyLog {
+            mango_group: *mango_group_ai.key,
+            liqee: *liqee_mango_account_ai.key,
+            liqor: *liqor_mango_account_ai.key,
+            liab_index: liab_index as u64,
+            insurance_transfer,
+            socialized_loss: socialized_loss.to_bits(),
+            percentage_loss: percentage_loss.to_bits(),
+            cache_deposit_index: mango_cache.root_bank_cache[liab_index].deposit_index.to_bits()
+        });
 
         Ok(())
     }
@@ -3466,18 +3467,12 @@ impl Processor {
             last_update: now_ts,
         };
 
-        msg!(
-            "update_root_bank details: {{ \
-            \"mango_group_pk\": \"{}\", \
-            \"token_index\": {}, \
-            \"deposit_index\": {}, \
-            \"borrow_index\": {} \
-        }}",
-            mango_group_ai.key,
-            index,
-            root_bank.deposit_index.to_num::<f64>(),
-            root_bank.borrow_index.to_num::<f64>(),
-        );
+        mango_emit!(UpdateRootBankLog {
+            mango_group: *mango_group_ai.key,
+            token_index: index as u64,
+            deposit_index: mango_cache.root_bank_cache[index].deposit_index.to_bits(),
+            borrow_index: mango_cache.root_bank_cache[index].borrow_index.to_bits()
+        });
 
         Ok(())
     }
@@ -3489,8 +3484,8 @@ impl Processor {
         accounts: &[AccountInfo],
         limit: usize,
     ) -> MangoResult<()> {
-        // Limit may be max 5 because of compute limits from logging. Increase if compute goes up
-        let limit = min(limit, 4);
+        // Limit may be max 10 because of compute limits from logging. Increase if compute goes up
+        let limit = min(limit, 10);
 
         const NUM_FIXED: usize = 4;
         let (fixed_ais, mango_account_ais) = array_refs![accounts, NUM_FIXED; ..;];
@@ -3540,6 +3535,7 @@ impl Processor {
                                 mango_group_ai.key,
                             )?,
                         };
+                        let pre_mngo = ma.perp_accounts[market_index].mngo_accrued;
                         ma.execute_maker(
                             market_index,
                             &mut perp_market,
@@ -3554,6 +3550,12 @@ impl Processor {
                             perp_market_cache,
                             fill,
                         )?;
+                        mango_emit!(MngoAccrualLog {
+                            mango_group: *mango_group_ai.key,
+                            mango_account: fill.maker,
+                            market_index: market_index as u64,
+                            mngo_accrual: ma.perp_accounts[market_index].mngo_accrued - pre_mngo
+                        });
                     } else {
                         let mut maker =
                             match mango_account_ais.iter().find(|ai| ai.key == &fill.maker) {
@@ -3579,6 +3581,7 @@ impl Processor {
                                     mango_group_ai.key,
                                 )?,
                             };
+                        let pre_mngo = maker.perp_accounts[market_index].mngo_accrued;
 
                         maker.execute_maker(
                             market_index,
@@ -3594,40 +3597,14 @@ impl Processor {
                             perp_market_cache,
                             fill,
                         )?;
+                        mango_emit!(MngoAccrualLog {
+                            mango_group: *mango_group_ai.key,
+                            mango_account: fill.maker,
+                            market_index: market_index as u64,
+                            mngo_accrual: maker.perp_accounts[market_index].mngo_accrued - pre_mngo
+                        });
                     }
-
-                    // TODO OPT remove this log if we start hitting compute limits
-                    // sol_log_compute_units();
-                    // msg!(
-                    //     "FillEvent details: {{ \
-                    //         \"timestamp\": {}, \
-                    //         \"seq_num\": {}, \
-                    //         \"maker\": {}, \
-                    //         \"taker\": {}, \
-                    //         \"taker_side\": {}, \
-                    //         \"maker_order_id\": {}, \
-                    //         \"taker_order_id\": {}, \
-                    //         \"maker_fee\": {}, \
-                    //         \"taker_fee\": {}, \
-                    //         \"price\": {}, \
-                    //         \"quantity\": {} \
-                    //         }}",
-                    //     fill.timestamp,
-                    //     fill.seq_num,
-                    //     fill.maker.to_string(),
-                    //     fill.taker.to_string(),
-                    //     if fill.taker_side == Side::Bid { "bid" } else { "sell" },
-                    //     fill.maker_order_id,
-                    //     fill.taker_order_id,
-                    //     fill.maker_fee.to_num::<f64>(),
-                    //     fill.taker_fee.to_num::<f64>(),
-                    //     fill.price,
-                    //     fill.quantity
-                    // );
-                    sol_log_compute_units();
-                    msg!(base64::encode(LoggableFillEvent::from_fill(fill).try_to_vec().unwrap())
-                        .as_str());
-                    sol_log_compute_units();
+                    mango_emit!(fill.to_fill_log(*mango_group_ai.key));
                 }
                 EventType::Out => {
                     let out: &OutEvent = cast_ref(event);
@@ -3968,7 +3945,6 @@ impl Processor {
         check!(price.is_positive(), MangoErrorCode::InvalidParam)?;
         check!(quantity.is_positive(), MangoErrorCode::InvalidParam)?;
         check!(trigger_price.is_positive(), MangoErrorCode::InvalidParam)?; // Is this necessary?
-        check!(reduce_only, MangoErrorCode::InvalidParam)?; // necessary restriction for now
 
         const NUM_FIXED: usize = 7;
         // TODO - *** amend client functions
@@ -4210,18 +4186,6 @@ impl Processor {
         let mut event_queue =
             EventQueue::load_mut_checked(event_queue_ai, program_id, &perp_market)?;
 
-        // We know all orders are reduce only
-        // What happens when there's a StopLimit order. that order wouldn't get placed in normal perp order
-        // If limit order and placing it puts you below Init, then remove
-
-        /*
-           what if you execute the order against another account you own when orderbook is empty?
-               normally that order wouldn't succeed, but here it would
-
-           soln: limit order should be at least as much as what we want
-
-        */
-
         // If reduce_only, position must only go down
         let quantity = if order.reduce_only {
             let base_pos = mango_account.perp_accounts[market_index].base_position;
@@ -4256,6 +4220,16 @@ impl Processor {
             )?;
 
             if sim_post_health >= ZERO_I80F48 || (health_up_only && sim_post_health >= pre_health) {
+                let (taker_base, taker_quote, bids_quantity, asks_quantity) = {
+                    let pa = &mango_account.perp_accounts[market_index];
+                    (
+                        pa.taker_base + taker_base,
+                        pa.taker_quote + taker_quote,
+                        pa.bids_quantity + bids_quantity,
+                        pa.asks_quantity + asks_quantity,
+                    )
+                };
+
                 book.new_order(
                     &mut event_queue,
                     &mut perp_market,
@@ -4280,6 +4254,7 @@ impl Processor {
                 )?;
                 let post_health = health_cache.get_health(&mango_group, HealthType::Init);
                 let pa = &mango_account.perp_accounts[market_index];
+
                 msg!("{:?} {:?}", sim_post_health, post_health);
                 msg!(
                     "{} {} {} {} {} {} {} {}",
@@ -4906,25 +4881,17 @@ fn checked_change_net(
 ) -> MangoResult<()> {
     if native_quantity.is_negative() {
         checked_sub_net(root_bank_cache, node_bank, mango_account, token_index, -native_quantity)?;
-        msg!(
-            "checked_sub_net details: {{ \"mango_group_pk\": \"{}\", \"mango_account_pk\": \"{}\", \"token_index\": {}, \"deposit\": {}, \"borrow\": {} }}",
-            mango_account.mango_group,
-            mango_account_pk, 
-            token_index,
-            mango_account.deposits[token_index].to_num::<f64>(),
-            mango_account.borrows[token_index].to_num::<f64>(),
-        );
     } else if native_quantity.is_positive() {
         checked_add_net(root_bank_cache, node_bank, mango_account, token_index, native_quantity)?;
-        msg!(
-            "checked_add_net details: {{ \"mango_group_pk\": \"{}\", \"mango_account_pk\": \"{}\", \"token_index\": {}, \"deposit\": {}, \"borrow\": {} }}",
-            mango_account.mango_group,
-            mango_account_pk,
-            token_index,
-            mango_account.deposits[token_index].to_num::<f64>(),
-            mango_account.borrows[token_index].to_num::<f64>(),
-        );
     }
+    mango_emit!(TokenBalanceLog {
+        mango_group: mango_account.mango_group,
+        mango_account: *mango_account_pk,
+        token_index: token_index as u64,
+        deposit: mango_account.deposits[token_index].to_bits(),
+        borrow: mango_account.borrows[token_index].to_bits()
+    });
+
     Ok(()) // This is an optimization to prevent unnecessary I80F48 calculations
 }
 
@@ -5238,6 +5205,9 @@ fn invoke_init_open_orders<'a>(
     solana_program::program::invoke_signed(&instruction, &account_infos, signers_seeds)
 }
 
+/*
+TODO test order types
+ */
 fn invoke_transfer_lamports<'a>(
     src_ai: &AccountInfo<'a>,
     dst_ai: &AccountInfo<'a>,
@@ -5299,12 +5269,10 @@ fn program_transfer_lamports(
     dst_ai: &AccountInfo,
     quantity: u64,
 ) -> MangoResult<()> {
-    let src_lamports =
-        src_ai.lamports().checked_sub(quantity).ok_or(throw_err!(MangoErrorCode::MathError))?;
+    let src_lamports = src_ai.lamports().checked_sub(quantity).ok_or(math_err!())?;
     **src_ai.lamports.borrow_mut() = src_lamports;
 
-    let dst_lamports =
-        dst_ai.lamports().checked_add(quantity).ok_or(throw_err!(MangoErrorCode::MathError))?;
+    let dst_lamports = dst_ai.lamports().checked_add(quantity).ok_or(math_err!())?;
     **dst_ai.lamports.borrow_mut() = dst_lamports;
     Ok(())
 }
@@ -5322,4 +5290,12 @@ fn cancel_all_advanced_orders<'a>(
         }
     }
     program_transfer_lamports(advanced_orders_ai, agent_ai, total_fee)
+}
+
+// Returns asset_weight and liab_weight
+fn get_leverage_weights(leverage: I80F48) -> (I80F48, I80F48) {
+    (
+        (leverage - ONE_I80F48).checked_div(leverage).unwrap(),
+        (leverage + ONE_I80F48).checked_div(leverage).unwrap(),
+    )
 }